requirements.txt
django>=4.2,<5.0
markdown>=3.5
pyyaml>=6.0
django-allauth>=0.61
python-dotenv>=1.0
bleach>=6.0
